requirements.txt
pandas==2.3.3
scikit-learn==1.6.1
kagglehub==0.4.0
imbalanced-learn==0.14.1
seaborn==0.13.2
matplotlib==3.10.8
pydantic==2.12.5
fastapi==0.128.0
uvicorn==0.40.0
pygad==2.19.0
cloudpickle==3.1.2
python-dotenv==1.0.0
setuptools>=65.0.0
requests>=2.31.0
crewai==0.80.0